+++ src/views/AddPriceSingle.vue
@@ -4,56 +4,19 @@
   <v-form @submit.prevent="createPrice">
     <v-row>
 
-      <!-- Step 1: proof -->
+      <!-- Step 1: product -->
       <v-col cols="12" md="6" lg="4">
         <v-card
-          title="Take a picture of the price tag"
-          subtitle="We need this for proof"
-          :prepend-icon="proofFormFilled ? 'mdi-image-check' : 'mdi-camera'"
+          title="Product info"
+          :prepend-icon="productFormFilled ? 'mdi-database-check-outline' : 'mdi-database-outline'"
           height="100%"
-          :style="proofFormFilled ? 'border: 1px solid #4CAF50' : 'border: 1px solid transparent'">
+          :style="productFormFilled ? 'border: 1px solid #4CAF50' : 'border: 1px solid transparent'">
           <v-divider></v-divider>
           <v-card-text>
-            <v-row>
-              <v-col>
-                <v-btn class="mb-2" size="small" prepend-icon="mdi-plus" @click.prevent="$refs.proof.click()" :loading="createProofLoading" :disabled="createProofLoading">Proof</v-btn>
-                <v-file-input
-                  class="overflow-hidden d-none"
-                  ref="proof"
-                  :prepend-icon="proofFormFilled ? 'mdi-image-check' : 'mdi-camera'"
-                  v-model="proofImage"
-                  capture="environment"
-                  accept="image/*"
-                  @change="uploadProof"
-                  @click:clear="clearProof"
-                  :loading="createProofLoading">
-                </v-file-input>
-                <p v-if="proofFormFilled && !createProofLoading" class="text-green mb-2"><i>Proof uploaded!</i></p>
-                <p v-if="!proofFormFilled && !createProofLoading" class="text-red mb-2"><i>Upload a proof</i></p>
-              </v-col>
-              <v-col v-if="proofFormFilled">
-                <v-img :src="proofImagePreview" style="max-height:200px"></v-img>
-              </v-col>
-            </v-row>
-          </v-card-text>
-        </v-card>
-      </v-col>
-
-      <!-- Step 2: product & price -->
-      <v-col cols="12" md="6" lg="4">
-        <v-card
-          title="Product & price details"
-          subtitle="The most important :)"
-          :prepend-icon="productPriceFormFilled ? 'mdi-tag-check-outline' : 'mdi-tag-outline'"
-          height="100%"
-          :style="productPriceFormFilled ? 'border: 1px solid #4CAF50' : 'border: 1px solid transparent'">
-          <v-divider></v-divider>
-          <v-card-text>
-            <h3 class="mb-1">
-              Product
+            <h3 class="mb-2">
               <v-item-group v-model="productMode" class="d-inline" mandatory>
                 <v-item v-for="pm in productModeList" :key="pm.key" :value="pm.key" v-slot="{ isSelected, toggle }">
-                  <v-chip class="mr-1" @click="toggle">
+                  <v-chip class="mr-1" @click="toggle" :style="isSelected ? 'border: 1px solid #9E9E9E' : 'border: 1px solid transparent'">
                     <v-icon start :icon="isSelected ? 'mdi-checkbox-marked-circle' : 'mdi-circle-outline'"></v-icon>
                     {{ pm.value }}
                   </v-chip>
@@ -61,7 +24,7 @@
               </v-item-group>
             </h3>
             <v-sheet v-if="productMode === 'barcode'">
-              <v-btn class="mb-2" size="small" prepend-icon="mdi-plus" @click="showBarcodeScanner">Scan a barcode</v-btn>
+              <v-btn class="mb-2" size="small" prepend-icon="mdi-barcode-scan" @click="showBarcodeScanner">Scan a barcode</v-btn>
               <v-text-field
                 v-if="dev"
                 :prepend-inner-icon="productBarcodeFormFilled ? 'mdi-barcode' : 'mdi-barcode-scan'"
@@ -69,9 +32,10 @@
                 label="Product code"
                 type="text"
                 hint="EAN"
+                hide-details="auto"
                 @click:prepend="showBarcodeScanner"
               ></v-text-field>
-              <PriceCard v-if="product" class="mb-4" :product="product" :readonly="true" elevation="1"></PriceCard>
+              <ProductCard v-if="product" class="mb-4" :product="product" :readonly="true" elevation="1"></ProductCard>
             </v-sheet>
             <v-sheet v-if="productMode === 'category'">
               <v-row>
@@ -102,22 +66,50 @@
               </div>
             </v-sheet>
             <p v-if="(productMode === 'barcode' && !productBarcodeFormFilled) || (productMode === 'category' && !productCategoryFormFilled)" class="text-red mb-2"><i>Set a product</i></p>
+          </v-card-text>
+        </v-card>
+      </v-col>
 
+      <!-- Step 2: price & proof -->
+      <v-col cols="12" md="6" lg="4">
+        <v-card
+          title="Price details"
+          :prepend-icon="priceProofFormFilled ? 'mdi-tag-check-outline' : 'mdi-tag-outline'"
+          height="100%"
+          :style="priceProofFormFilled ? 'border: 1px solid #4CAF50' : 'border: 1px solid transparent'">
+          <v-divider></v-divider>
+          <v-card-text>
             <h3 class="mb-1">Price <span v-if="productMode === 'category'">per kg</span></h3>
             <v-row>
-              <v-col cols="6">
+              <v-col cols="12" sm="6">
                 <v-text-field
                   v-model="addPriceSingleForm.price"
                   label="Price"
                   type="number"
+                  hide-details="auto"
+                  :suffix="addPriceSingleForm.currency"
                 ></v-text-field>
               </v-col>
-              <v-col cols="6">
-                <v-autocomplete
-                  v-model="addPriceSingleForm.currency"
-                  label="Currency"
-                  :items="currencyList"
-                ></v-autocomplete>
+            </v-row>
+            <v-row>
+              <v-col>
+                <v-btn class="mb-2" size="small" prepend-icon="mdi-camera" @click.prevent="$refs.proof.click()" :loading="createProofLoading" :disabled="createProofLoading">Proof</v-btn>
+                <v-file-input
+                  class="overflow-hidden d-none"
+                  ref="proof"
+                  :prepend-icon="proofFormFilled ? 'mdi-image-check' : 'mdi-camera'"
+                  v-model="proofImage"
+                  capture="environment"
+                  accept="image/*"
+                  @change="uploadProof"
+                  @click:clear="clearProof"
+                  :loading="createProofLoading">
+                </v-file-input>
+                <p v-if="proofFormFilled && !createProofLoading" class="text-green mb-2"><i>Proof uploaded!</i></p>
+                <p v-if="!proofFormFilled && !createProofLoading" class="text-red mb-2"><i>Upload a proof</i></p>
+              </v-col>
+              <v-col v-if="proofFormFilled">
+                <v-img :src="proofImagePreview" style="max-height:200px"></v-img>
               </v-col>
             </v-row>
           </v-card-text>
@@ -128,7 +120,6 @@
       <v-col cols="12" md="6" lg="4">
         <v-card
           title="Where & when?"
-          subtitle="Final step!"
           :prepend-icon="locationDateFormFilled ? 'mdi-map-marker-check-outline' : 'mdi-map-marker-outline'"
           height="100%"
           :style="locationDateFormFilled ? 'border: 1px solid #4CAF50' : 'border: 1px solid transparent'">
@@ -137,20 +128,21 @@
             <h3 class="mb-1">
               Location
             </h3>
-            <v-btn class="mb-2" size="small" prepend-icon="mdi-plus" @click="showLocationSelector">Find</v-btn>
             <v-chip
               class="mb-2"
-              :style="isSelectedLocation(location) ? 'border: 1px solid #4CAF50' : 'border: 1px solid transparent'"
+              :style="isSelectedLocation(location) ? 'border: 1px solid #9E9E9E' : 'border: 1px solid transparent'"
               v-for="location in recentLocations"
               @click="setLocationData(location)">
               <v-icon start :icon="isSelectedLocation(location) ? 'mdi-checkbox-marked-circle' : 'mdi-history'"></v-icon>
-              {{ location.display_name }}
+              {{ getNominatimLocationTitle(location, true, true, true) }}
             </v-chip>
+            <br v-if="recentLocations.length" />
+            <v-btn class="mb-2" size="small" prepend-icon="mdi-magnify" @click="showLocationSelector">Find</v-btn>
             <p v-if="!locationFormFilled" class="text-red mb-2"><i>Select your location</i></p>
 
             <h3 class="mt-4 mb-1">Date</h3>
             <v-row>
-              <v-col>
+              <v-col cols="12" sm="6">
                 <v-text-field
                   v-model="addPriceSingleForm.date"
                   label="Date"
@@ -195,9 +187,9 @@
 import Compressor from 'compressorjs'
 import { mapStores } from 'pinia'
 import { useAppStore } from '../store'
-import constants from '../constants'
 import api from '../services/api'
-import PriceCard from '../components/PriceCard.vue'
+import utils from '../utils.js'
+import ProductCard from '../components/ProductCard.vue'
 import BarcodeScanner from '../components/BarcodeScanner.vue'
 import LocationSelector from '../components/LocationSelector.vue'
 import CategoryTags from '../data/category-tags.json'
@@ -214,7 +206,7 @@ Compressor.setDefaults({
 
 export default {
   components: {
-    PriceCard,
+    ProductCard,
     BarcodeScanner,
     LocationSelector
   },
@@ -223,7 +215,6 @@ export default {
       dev: import.meta.env.DEV,
       // price form
       addPriceSingleForm: {
-        proof_id: null,
         product_code: '',
         category_tag: null,
         origins_tags: '',
@@ -232,14 +223,10 @@ export default {
         currency: null,  // see initPriceSingleForm
         location_osm_id: null,
         location_osm_type: '',
-        date: new Date().toISOString().substr(0, 10)
+        date: new Date().toISOString().substr(0, 10),
+        proof_id: null,
       },
       createPriceLoading: false,
-      // proof data
-      proofImage: null,
-      proofImagePreview: null,
-      createProofLoading: false,
-      proofSuccessMessage: false,
       // product data
       product: null,
       productModeList: [{key: 'barcode', value: 'Barcode', icon: 'mdi-barcode-scan'}, {key: 'category', value: 'Category', icon: 'mdi-basket-outline'}],
@@ -248,19 +235,18 @@ export default {
       originsTags: OriginsTags,  // list of origins tags for autocomplete
       labelsTags: LabelsTags,
       barcodeScanner: false,
-      // price data
-      currencyList: constants.CURRENCY_LIST,
       // location data
       locationSelector: false,
       locationSelectedDisplayName: '',
+      // proof data
+      proofImage: null,
+      proofImagePreview: null,
+      createProofLoading: false,
+      proofSuccessMessage: false,
     }
   },
   computed: {
     ...mapStores(useAppStore),
-    proofFormFilled() {
-      let keys = ['proof_id']
-      return Object.keys(this.addPriceSingleForm).filter(k => keys.includes(k)).every(k => !!this.addPriceSingleForm[k])
-    },
     productBarcodeFormFilled() {
       let keys = ['product_code']
       return Object.keys(this.addPriceSingleForm).filter(k => keys.includes(k)).every(k => !!this.addPriceSingleForm[k])
@@ -269,8 +255,16 @@ export default {
       let keys = ['category_tag', 'origins_tags']
       return Object.keys(this.addPriceSingleForm).filter(k => keys.includes(k)).every(k => !!this.addPriceSingleForm[k])
     },
-    productPriceFormFilled() {
-      return (this.productBarcodeFormFilled || this.productCategoryFormFilled) && !!this.addPriceSingleForm.price && !!this.addPriceSingleForm.currency
+    productFormFilled() {
+      return this.productBarcodeFormFilled || this.productCategoryFormFilled
+    },
+    priceProofFormFilled() {
+      let keys = ['price', 'currency', 'proof_id']
+      return Object.keys(this.addPriceSingleForm).filter(k => keys.includes(k)).every(k => !!this.addPriceSingleForm[k])
+    },
+    proofFormFilled() {
+      let keys = ['proof_id']
+      return Object.keys(this.addPriceSingleForm).filter(k => keys.includes(k)).every(k => !!this.addPriceSingleForm[k])
     },
     recentLocations() {
       return this.appStore.getRecentLocations(3)
@@ -284,7 +278,7 @@ export default {
       return Object.keys(this.addPriceSingleForm).filter(k => keys.includes(k)).every(k => !!this.addPriceSingleForm[k])
     },
     formFilled() {
-      return this.proofFormFilled && this.productPriceFormFilled && this.locationDateFormFilled
+      return this.productFormFilled && this.priceProofFormFilled && this.locationDateFormFilled
     },
   },
   mounted() {
@@ -295,8 +289,14 @@ export default {
       return !!v
     },
     initPriceSingleForm() {
+      /**
+       * init product mode, currency & last location
+       */
       this.productMode = this.appStore.user.last_product_mode_used
       this.addPriceSingleForm.currency = this.appStore.user.last_currency_used
+      if (this.recentLocations.length) {
+        this.setLocationData(this.recentLocations[0])
+      }
     },
     clearProof() {
       this.proofImage = null
@@ -378,12 +378,12 @@ export default {
       this.addPriceSingleForm.product_code = code
       this.product = null
       api
-        .openfoodfactsProductSearch(code)
+        .getProductByCode(code)
         .then((data) => {
-          this.product = data['product'] || {'code': code}
+          this.product = data.id ? data : {'code': code}
         })
         .catch((error) => {
-          alert("Error: Open Food Facts server error")
+          alert("Error: Open Prices server error")
         })
     },
     showLocationSelector() {
@@ -391,6 +391,9 @@ export default {
     },
     closeLocationSelector(event) {
       this.locationSelector = false
+    },
+    getNominatimLocationTitle(location, withName=true, withRoad=false, withCity=true) {
+      return utils.getLocationTitle(location, withName, withRoad, withCity)
     },
     setLocationData(location) {
       this.appStore.addRecentLocation(location)