+++ src/views/Stats.vue
@@ -6,7 +6,10 @@
 
   <v-row>
     <v-col cols="12" md="6" lg="4">
-      <v-card title="Price count" :text="priceTotal"></v-card>
+      <v-card title="Prices" :text="priceTotal" height="100%"></v-card>
+    </v-col>
+    <v-col cols="12" md="6" lg="4">
+      <v-card title="Products with prices" :subtitle="'Out of ' + productTotal + ' products'" :text="productWithPriceTotal"></v-card>
     </v-col>
   </v-row>
 </template>
@@ -18,11 +21,15 @@ export default {
   data() {
     return {
       priceTotal: null,
+      productTotal: null,
+      productWithPriceTotal: null,
       loading: false,
     }
   },
   mounted() {
     this.getPrices()
+    this.getProducts()
+    this.getProductsWithPrice()
   },
   methods: {
     getPrices() {
@@ -30,6 +37,22 @@ export default {
       return api.getPrices({ size: 1 })
         .then((data) => {
           this.priceTotal = data.total
+          this.loading = false
+        })
+    },
+    getProducts() {
+      this.loading = true
+      return api.getProducts({ size: 1 })
+        .then((data) => {
+          this.productTotal = data.total
+          this.loading = false
+        })
+    },
+    getProductsWithPrice() {
+      this.loading = true
+      return api.getProducts({ size: 1, price_count__gte: 1 })
+        .then((data) => {
+          this.productWithPriceTotal = data.total
           this.loading = false
         })
     }