+++ src/views/UserSettings.vue
@@ -1,0 +1,64 @@
+<template>
+  <h1 class="mb-1">
+    Settings
+  </h1>
+
+  <v-form @submit.prevent="updateSettings">
+    <v-row>
+      <v-col cols="12" sm="6">
+        <v-card title="Adding prices" prepend-icon="mdi-tag-plus-outline">
+          <v-divider></v-divider>
+          <v-card-text>
+            <h3 class="mb-1">Default currency</h3>
+            <v-autocomplete
+              v-model="userSettingsForm.currency"
+              label="Currency"
+              :items="currencyList"
+            ></v-autocomplete>
+          </v-card-text>
+        </v-card>
+      </v-col>
+    </v-row>
+
+    <v-row>
+      <v-col>
+        <v-btn type="submit" :color="formFilled ? 'success' : ''" :disabled="!formFilled">Save</v-btn>
+      </v-col>
+    </v-row>
+  </v-form>
+</template>
+
+<script>
+import { mapStores } from 'pinia'
+import { useAppStore } from '../store'
+import constants from '../constants'
+
+export default {
+  data() {
+    return {
+      userSettingsForm: {
+        currency: null,  // see initUserSettingsForm
+      },
+      currencyList: constants.CURRENCY_LIST,
+    }
+  },
+  computed: {
+    ...mapStores(useAppStore),
+    formFilled() {
+      return Object.values(this.userSettingsForm).every(x => !!x)
+    }
+  },
+  mounted() {
+    this.initUserSettingsForm()
+  },
+  methods: {
+    initUserSettingsForm() {
+      this.userSettingsForm.currency = this.appStore.user.last_currency_used
+    },
+    updateSettings() {
+      this.appStore.setLastCurrencyUsed(this.userSettingsForm.currency)
+      this.$router.push({ path: '/', query: { settingsSuccess: 'true' } })
+    }
+  }
+}
+</script>